fix: move media query ranges into core
