+++ packages/core/src/createStringify.js
@@ -8,6 +8,8 @@ const captureTokens = /([+-])?((?:\d+(?:\.\d*)?|\.\d+)(?:[Ee][+-]?\d+)?)?(\$|--)
 const splitBySpace = /\s+(?![^()]*\))/
 const split = (fn) => (data) => fn(...(typeof data === 'string' ? String(data).split(splitBySpace) : [data]))
 
+const mqunit = /([\d.]+)([^]*)/
+
 const patches = {
 	// prefixed properties
 	'appearance': (d) => ({ WebkitAppearance: d, appearance: d }),
@@ -56,7 +58,39 @@ export const createStringify = (config) => {
 			const firstChar = name.charCodeAt(0)
 
 			/** CSS property name, which may be a specially-formatted custom property. */
-			const customName = firstChar === 36 ? '-' + name.replace(/\$/g, '-') : name
+			let customName =
+				// prettier-ignore
+				firstChar === 64
+					? (
+						name.slice(1) in media
+							? '@media ' + media[name.slice(1)]
+						: name
+					).replace(/\(\s*([\w-]+)\s*(=|<|<=|>|>=)\s*([\w-]+)\s*(?:(<|<=|>|>=)\s*([\w-]+)\s*)?\)/g, (_, a, l, b, r, c) => {
+						const isValueFirst = mqunit.test(a)
+						const shift = 0.0625 * (isValueFirst ? -1 : 1)
+						const [name, value] = isValueFirst ? [b, a] : [a, b]
+
+						return (
+							// prettier-ignore
+							'(' +
+								(
+									l[0] === '=' ? '' : (l[0] === '>' === isValueFirst ? 'max-' : 'min-')
+								) + name + ':' +
+								(l[0] !== '=' && l.length === 1 ? value.replace(mqunit, (_, v, u) => Number(v) + shift * (l === '>' ? 1 : -1) + u) : value) +
+								(
+									r
+										? ') and (' + (
+											(r[0] === '>' ? 'min-' : 'max-') + name + ':' +
+											(r.length === 1 ? c.replace(mqunit, (_, v, u) => Number(v) + shift * (r === '>' ? -1 : 1) + u) : c)
+										)
+									: ''
+								) +
+							')'
+						)
+					})
+				: firstChar === 36
+					? '-' + name.replace(/\$/g, '-')
+				: name
 
 			const customData =
 				// prettier-ignore
@@ -102,12 +136,6 @@ export const createStringify = (config) => {
 			if (data !== customData || name !== customName) {
 				return {
 					[customName]: customData,
-				}
-			}
-
-			if (firstChar === 64 && name.slice(1) in media) {
-				return {
-					['@media ' + media[name.slice(1)]]: data,
 				}
 			}
